unit_6/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.2.0
python-dotenv==1.0.0
avro-python3==1.10.2